test: update logger tests
